Refactor: refactor names, and adjustments

+++ _src/domain/usecases/product/DeleteProductUseCase.ts
@@ -1,9 +1,10 @@
 import { IProductRepository } from "../../repositories/IProductRepository";
 
-export class DeleteProducts {
+export class DeleteProductsUseCase {
     constructor(private productRepository: IProductRepository) {};
 
     async execute(id: string) :Promise<void> {
-        return this.productRepository.deleteProduct(id);
+        if(!id) throw new Error('Id is required to delete');
+        await this.productRepository.deleteProduct(id);
     }
 }
